Clear the -Wall compile warnings (portable, behaviour-preserving)
A clean `make` now builds with no warnings. All fixes are byte-identical --
they either make an implicit conversion explicit or remove dead code:

- -Wimplicit-const-int-float-conversion: `-INFI` in the score sums is an int
  (2147483647) the compiler was silently rounding to float; make it
  `-(float)INFI` (same value, INFI stays an int for its long "infinity"
  sentinel uses in ReadGeneModel/genamic/BackupGenes). beggar.c's HSP byte
  count is cast to float (it is only a rough -B estimate).
- -Wabsolute-value: the branch-point penalty called abs() on a long arg
  (int param -> truncation risk); use labs() (BuildAcceptors/BuildU12Acceptors).
- -Wformat-security: printHelp did printf(USAGE) with a non-literal format;
  use printf("%s", USAGE) (USAGE has no % specifiers, so output is identical).
- -Wparentheses-equality: drop the extra parens in `if ((E->Strand == '*'))`.
- -Wunused-but-set-variable: remove dead locals featnum (CookingGenes), rss
  (ScoreExons), line9 (ReadHSP; it was the optional 9th HSP column, the last
  strtok on the line, never read), and LowestLocalExon/js (BuildZeroLengthExons).
- -Wmisleading-indentation: reindent the UTR `if` in CookingInfo and the Ga
  allocation loop in RequestMemoryGenes.

Some of these are clang-specific warning flags, but the fixes (explicit
casts, labs, %s) are good practice on gcc too and introduce no new warnings.
No identifiers renamed. Regression suite 5/5 byte-identical.

+++ src/GetStopCodons.c
@@ -73,7 +73,7 @@ long GetStopCodons(char* s,
 			  index = OligoToInt(s+j, p->order+1,5);
 			  
 			  if (index >= p->dimensionTrans)
-				score = score + -INFI;
+				score = score + -(float)INFI;
 			  else
 				score = score + p->transitionValues[i][index];
 			}
@@ -116,7 +116,7 @@ long GetStopCodons(char* s,
 				  /* i is the position inside the region */
 				  index = TRANS[(int)(*(s + i))];
 				  if (index >= p->dimensionTrans)
-					score = score + -INFI;
+					score = score + -(float)INFI;
 				  else
 					score = score + p->transitionValues[i][index];
 				}
@@ -155,7 +155,7 @@ long GetStopCodons(char* s,
 				  /* i is the position inside the region */
 				  index = 5*TRANS[(int)(*(s + i -1))] + TRANS[(int)(*(s + i))];
 				  if (index >= p->dimensionTrans)
-					score = score + -INFI;
+					score = score + -(float)INFI;
 				  else
 					score = score + p->transitionValues[i][index];
 				}
@@ -194,7 +194,7 @@ long GetStopCodons(char* s,
 				  /* i is the position inside the region */
 				  index = OligoToInt(s + i - p->order , p->order+1,5); 
 				  if (index >= p->dimensionTrans)
-					score = score + -INFI;
+					score = score + -(float)INFI;
 				  else
 					score = score + p->transitionValues[i][index];
 				}
